Bucky.send accepts type 'string'
Previously it was only possible to send numeric values (types: `timer`,
`gauge` and `counter`). This commit adds supports for a type `string`
which accepts arbitrary strings. The identifier in a measurement is `s`,
as in `this.is.a.string:hello|s`.

My current use-case requires this in order to pass a correlation ID
together with the measurements. Sharing as PR in hopes for it to be
useful to somebody.